forge hates me pt 2
