+++ forge/src/main/java/dev/mayaqq/estrogen/forge/client/EstrogenForgeClient.java
@@ -3,10 +3,13 @@ package dev.mayaqq.estrogen.forge.client;
 import com.simibubi.create.foundation.config.ui.BaseConfigScreen;
 import dev.mayaqq.estrogen.Estrogen;
 import dev.mayaqq.estrogen.client.EstrogenClient;
+import dev.mayaqq.estrogen.client.registry.blockRenderers.dreamBlock.DreamBlockShader;
 import dev.mayaqq.estrogen.resources.BreastArmorDataLoader;
+import net.minecraft.client.renderer.ShaderInstance;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.client.ConfigScreenHandler;
 import net.minecraftforge.client.event.RegisterClientReloadListenersEvent;
+import net.minecraftforge.client.event.RegisterShadersEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
@@ -28,5 +31,13 @@ public class EstrogenForgeClient {
     @SubscribeEvent
     public static void registerClientReloadListeners(RegisterClientReloadListenersEvent event) {
         event.registerReloadListener(BreastArmorDataLoader.INSTANCE);
+    }
+
+    @SubscribeEvent
+    public static void registerShaders(RegisterShadersEvent event) {
+        DreamBlockShader.register((id, format, shaderConsumer) -> {
+            ShaderInstance instance = new ShaderInstance(event.getResourceProvider(), id, format);
+            event.registerShader(instance, shaderConsumer);
+        });
     }
 }
